pyproject.toml
[tool.poetry]
name = "asynclient"
version = "0.1.0"
description = ""
authors = ["Sergey Yavorsky <yavorskyserge@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
telethon-repl = "^1.29.3"
pytube = "^15.0.0"
cryptg = "^0.4.0"
python-dotenv = "^1.0.0"
aiohttp = "^3.9.1"
aiojobs = "^1.2.1"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
